CCTVLIVE/Vijay-Vehicle_Accident_System/requirements.txt
torch
torchvision
transformers
numpy
Pillow
opencv-python
